삭제 하기 전에 삭제할 지 confirm

+++ src/webapp/jsp/usr/article/list.jsp
@@ -56,7 +56,7 @@ List<Article> articles = (List<Article>) request.getAttribute("articles");
           <td>
             <a href="/usr/article/modify/free/<%=article.getId()%>" class="badge badge-primary badge-outline">수정</a>
             &nbsp;
-            <a href="/usr/article/delete/free/<%=article.getId()%>" class="badge badge-secondary badge-outline">삭제</a>
+            <a onclick="if(!confirm('정말 삭제하시겠습니까?')) return false;" href="/usr/article/delete/free/<%=article.getId()%>" class="badge badge-secondary badge-outline">삭제</a>
           </td>
         </tr>
         <% } %>
